# Faster examples with an extended TMU usage