It turns out that the TMU can read/write at most four vectors at a time, and can prefetch into TMU cache without an explicit wait (see [nir_to_vir.c](https://gitlab.freedesktop.org/mesa/mesa/-/blob/main/src/broadcom/compiler/nir_to_vir.c) in Mesa for the usage), which increased the memory performance by a few percent:
```
summation: 7010 MB/s -> 7242 MB/s (+3.3%)
memset: 3739 MB/s -> 3752 MB/s (+0.35%)
scopy: 2386 MB/s -> 2424 MB/s (+1.6%)
```