@Terminus-IMRC 2021-06-30
@notogawa Added some tests, so can you review this pull request again please?